move toggle thema ready for multiple css themes
https://hamoria.atlassian.net/browse/HM-115
